@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
     <DockerComposeProjectPath>..\docker-compose.dcproj</DockerComposeProjectPath>
   </PropertyGroup>
 
@@ -10,26 +10,35 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="AutoMapper" Version="7.0.1" />
-    <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="5.0.1" />
-    <PackageReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="NLog" Version="4.5.10" />
-    <PackageReference Include="NLog.Extensions.Logging" Version="1.3.0" />
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="3.0.0" />
+    <PackageReference Include="AutoMapper" Version="9.0.0" />
+    <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="7.0.0" />
+    <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="3.0.0" />
+    <PackageReference Include="NLog" Version="4.6.8" />
+    <PackageReference Include="NLog.Extensions.Logging" Version="1.6.1" />
+    <PackageReference Include="Swashbuckle.AspNetCore" Version="5.0.0-rc4" />
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\ASPNetCore.CleanArchitecture.Exceptions\ASPNetCore.CleanArchitecture.Exceptions.csproj">
+      <Name>ASPNetCore.CleanArchitecture.Exceptions</Name>
+    </ProjectReference>
     <ProjectReference Include="..\ASPNetCore.CleanArchitecture.Interfaces\ASPNetCore.CleanArchitecture.Interfaces.csproj">
       <Name>ASPNetCore.CleanArchitecture.Interfaces</Name>
     </ProjectReference>
     <ProjectReference Include="..\ASPNetCore.CleanArchitecture.Models\ASPNetCore.CleanArchitecture.Models.csproj">
       <Name>ASPNetCore.CleanArchitecture.Models</Name>
     </ProjectReference>
+    <ProjectReference Include="..\ASPNetCore.CleanArchitecture.Patterns\ASPNetCore.CleanArchitecture.Patterns.csproj">
+      <Name>ASPNetCore.CleanArchitecture.Patterns</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\ASPNetCore.CleanArchitecture.Resources\ASPNetCore.CleanArchitecture.Resources.csproj">
+      <Name>ASPNetCore.CleanArchitecture.Resources</Name>
+    </ProjectReference>
     <ProjectReference Include="..\ASPNetCore.CleanArchitecture.Setup\ASPNetCore.CleanArchitecture.Setup.csproj">
       <Name>ASPNetCore.CleanArchitecture.Setup</Name>
     </ProjectReference>
   </ItemGroup>
-
+  
   <PropertyGroup>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
     <NoWarn>$(NoWarn);1591</NoWarn>
