@@ -1,23 +1,25 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="AutoMapper" Version="7.0.1" />
+    <PackageReference Include="AutoMapper" Version="9.0.0" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="3.0.1" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="3.0.1" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\ASPNetCore.CleanArchitecture.Data\ASPNetCore.CleanArchitecture.Data.csproj">
       <Name>ASPNetCore.CleanArchitecture.Data</Name>
     </ProjectReference>
-    <ProjectReference Include="..\ASPNetCore.CleanArchitecture.Models\ASPNetCore.CleanArchitecture.Models.csproj">
-      <Name>ASPNetCore.CleanArchitecture.Models</Name>
-    </ProjectReference>
     <ProjectReference Include="..\ASPNetCore.CleanArchitecture.Interfaces\ASPNetCore.CleanArchitecture.Interfaces.csproj">
       <Name>ASPNetCore.CleanArchitecture.Interfaces</Name>
     </ProjectReference>
+    <ProjectReference Include="..\ASPNetCore.CleanArchitecture.Models\ASPNetCore.CleanArchitecture.Models.csproj">
+      <Name>ASPNetCore.CleanArchitecture.Models</Name>
+    </ProjectReference>
   </ItemGroup>
-  
+
 </Project>
