@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
@@ -9,4 +9,5 @@
       <Name>ASPNetCore.CleanArchitecture.Models</Name>
     </ProjectReference>
   </ItemGroup>
+
 </Project>
