@@ -3,13 +3,11 @@
 /// https://github.com/medalinouira
 /// Copyright © Mohamed Ali NOUIRA. All rights reserved.
 
+using ASPNetCore.CleanArchitecture.Models;
 using AutoMapper;
-
 using ASPNetCore.CleanArchitecture.Api.ViewModels.Order;
 using ASPNetCore.CleanArchitecture.Api.ViewModels.Product;
 using ASPNetCore.CleanArchitecture.Api.ViewModels.Customer;
-
-using ASPNetCore.CleanArchitecture.Models;
 
 namespace ASPNetCore.CleanArchitecture.Api.Mapping
 {
